+++ src/components/people-search/Person/PersonInfo.js
